Preparing data for LIDIF function.

The function prepare your raw data to fit into LIDIF model.

Usage

prepare_data(X, Z)

Arguments

X

A numeric matrix with columns corresponding to items. If the item is binary, please code it as 0/1; if the item is ordinal, please code it as integers starting from 1, e.g. 1,2,3. X can be missing at some columns but not the entire row.

Z

the matrix of exogenous variables. Z should not contain any missing data and should have the same number of rows as X. Please do not include intercept here.

Value

A list containing two elements.

Examples

items_matrix = cbind.data.frame("item1" = rbinom(100, 0,1),
                                "item2" = sample(c(1,2,3), size = 100, replace = TRUE))
covariates = cbind.data.frame("age" = runif(100, 10, 50),
                             "sex" = rbinom(100, 0, 1))
prepare_data(X = items_matrix, Z = covariates)
#> $X
#>        item1 item2
#>   [1,]     0     1
#>   [2,]     0     3
#>   [3,]     0     3
#>   [4,]     0     1
#>   [5,]     0     2
#>   [6,]     0     3
#>   [7,]     0     3
#>   [8,]     0     3
#>   [9,]     0     2
#>  [10,]     0     1
#>  [11,]     0     1
#>  [12,]     0     1
#>  [13,]     0     3
#>  [14,]     0     3
#>  [15,]     0     1
#>  [16,]     0     2
#>  [17,]     0     2
#>  [18,]     0     3
#>  [19,]     0     3
#>  [20,]     0     2
#>  [21,]     0     3
#>  [22,]     0     1
#>  [23,]     0     1
#>  [24,]     0     1
#>  [25,]     0     2
#>  [26,]     0     2
#>  [27,]     0     3
#>  [28,]     0     3
#>  [29,]     0     2
#>  [30,]     0     3
#>  [31,]     0     1
#>  [32,]     0     2
#>  [33,]     0     3
#>  [34,]     0     3
#>  [35,]     0     1
#>  [36,]     0     2
#>  [37,]     0     1
#>  [38,]     0     2
#>  [39,]     0     3
#>  [40,]     0     2
#>  [41,]     0     1
#>  [42,]     0     2
#>  [43,]     0     1
#>  [44,]     0     2
#>  [45,]     0     3
#>  [46,]     0     1
#>  [47,]     0     2
#>  [48,]     0     2
#>  [49,]     0     3
#>  [50,]     0     2
#>  [51,]     0     1
#>  [52,]     0     1
#>  [53,]     0     3
#>  [54,]     0     2
#>  [55,]     0     1
#>  [56,]     0     1
#>  [57,]     0     3
#>  [58,]     0     3
#>  [59,]     0     2
#>  [60,]     0     3
#>  [61,]     0     2
#>  [62,]     0     3
#>  [63,]     0     2
#>  [64,]     0     3
#>  [65,]     0     2
#>  [66,]     0     2
#>  [67,]     0     1
#>  [68,]     0     3
#>  [69,]     0     2
#>  [70,]     0     2
#>  [71,]     0     3
#>  [72,]     0     1
#>  [73,]     0     2
#>  [74,]     0     1
#>  [75,]     0     1
#>  [76,]     0     3
#>  [77,]     0     2
#>  [78,]     0     1
#>  [79,]     0     2
#>  [80,]     0     2
#>  [81,]     0     2
#>  [82,]     0     3
#>  [83,]     0     2
#>  [84,]     0     2
#>  [85,]     0     2
#>  [86,]     0     1
#>  [87,]     0     3
#>  [88,]     0     3
#>  [89,]     0     1
#>  [90,]     0     1
#>  [91,]     0     1
#>  [92,]     0     1
#>  [93,]     0     2
#>  [94,]     0     1
#>  [95,]     0     1
#>  [96,]     0     1
#>  [97,]     0     2
#>  [98,]     0     1
#>  [99,]     0     3
#> [100,]     0     3
#> 
#> $Z
#>     Intercept      age sex
#> 1           1 19.94155   0
#> 2           1 26.11525   0
#> 3           1 40.78521   0
#> 4           1 14.77941   0
#> 5           1 17.78780   0
#> 6           1 16.58277   0
#> 7           1 36.52826   0
#> 8           1 44.26300   0
#> 9           1 47.06186   0
#> 10          1 32.09510   0
#> 11          1 33.08263   0
#> 12          1 37.49791   0
#> 13          1 19.78873   0
#> 14          1 11.78469   0
#> 15          1 46.39418   0
#> 16          1 12.82725   0
#> 17          1 49.87566   0
#> 18          1 34.47410   0
#> 19          1 16.90235   0
#> 20          1 46.37764   0
#> 21          1 11.49805   0
#> 22          1 33.74215   0
#> 23          1 19.47910   0
#> 24          1 46.25189   0
#> 25          1 42.75492   0
#> 26          1 37.99317   0
#> 27          1 18.80001   0
#> 28          1 39.11964   0
#> 29          1 18.68338   0
#> 30          1 28.24921   0
#> 31          1 23.31199   0
#> 32          1 32.73411   0
#> 33          1 20.08823   0
#> 34          1 28.56054   0
#> 35          1 46.70642   0
#> 36          1 48.91377   0
#> 37          1 42.76330   0
#> 38          1 46.11695   0
#> 39          1 33.25464   0
#> 40          1 40.92034   0
#> 41          1 49.80492   0
#> 42          1 38.43885   0
#> 43          1 18.59770   0
#> 44          1 21.67031   0
#> 45          1 38.87039   0
#> 46          1 44.66463   0
#> 47          1 19.53812   0
#> 48          1 10.17985   0
#> 49          1 47.74066   0
#> 50          1 27.52549   0
#> 51          1 40.02413   0
#> 52          1 36.71263   0
#> 53          1 26.31893   0
#> 54          1 24.04995   0
#> 55          1 39.52366   0
#> 56          1 36.57142   0
#> 57          1 13.40899   0
#> 58          1 44.24529   0
#> 59          1 13.07933   0
#> 60          1 44.11379   0
#> 61          1 14.25388   0
#> 62          1 29.39211   0
#> 63          1 19.88876   0
#> 64          1 37.46277   0
#> 65          1 16.54493   0
#> 66          1 48.11299   0
#> 67          1 22.87418   0
#> 68          1 24.46136   0
#> 69          1 45.50894   0
#> 70          1 43.12058   0
#> 71          1 14.02626   0
#> 72          1 46.24206   0
#> 73          1 40.90921   0
#> 74          1 25.33483   0
#> 75          1 49.98610   0
#> 76          1 23.97196   0
#> 77          1 47.89273   0
#> 78          1 18.64400   0
#> 79          1 11.28371   0
#> 80          1 15.81263   0
#> 81          1 44.17536   0
#> 82          1 18.52597   0
#> 83          1 18.41243   0
#> 84          1 11.58083   0
#> 85          1 47.79099   0
#> 86          1 19.79712   0
#> 87          1 41.24490   0
#> 88          1 21.52949   0
#> 89          1 45.01432   0
#> 90          1 21.83000   0
#> 91          1 49.34102   0
#> 92          1 33.59350   0
#> 93          1 40.36634   0
#> 94          1 43.44301   0
#> 95          1 40.51278   0
#> 96          1 26.69080   0
#> 97          1 15.52299   0
#> 98          1 13.23380   0
#> 99          1 36.23931   0
#> 100         1 34.08015   0
#>