Summarize LIDIF output.

Summarize LIDIF output.

Usage

summary_LIDIF(object, terms = NULL, all_terms = TRUE, digits = 2)

Arguments

object

An object of class `LIDIF`.

terms

A vector of characters containing the name of variables you would like to perform Wald tests. For example, specifying terms = "age" if you would like to test age's DIF effects. The default is `NULL`.

all_terms

Logical. Whether perform Wald tests on all individual coefficients. The default is `TRUE`.

digits

Numeric. Control the number of digits displayed in the results. The default value is 2.

Value

A list containing.

`stats`: the Wald statistics.

`p_value`: the corresponding p values.

`coefficients`: the data frame containing estimated coefficients, standard error, factor loading and p values.

`term`: the corresponding test statistics, degree of freedom and p values if specifying `terms` in the argument.

Examples

 if (FALSE) {
## load the sample data (see `?LIDIF::binsurvs` for details)
binsurvs = LIDIF::binsurvs # The exogenous variables are age and sex.

## prepare for model fitting
dat.list = prepare_data(binsurvs$X, binsurvs$Z)

## LIDIF fitting
# The parameter specified is set to reduce the running time.
# In your analysis, please refer to the default value.
library(LIDIF)
res = LIDIF(dat.list = dat.list,
            cl_num = 5, # before running, please check if you have 5 cores
            type_list = "binary",
            maxit = 2,
            random_per = 0.5,
            init_nums = 2,
            init_maxit = 2)

## Summarize results: individual DIF effects
summary_LIDIF(res)

## Summarize results: combineing uniform and non-uniform effects for sex
summary_LIDIF(res, terms = "sex")
}